Preserve leading zeros in #nnn issue links.
Sometimes a user would post an issue containing something like
"#01234" in the description/update.  If there happen to be an issue
with that number (and it is visible by the current user, which is
always true for site admin,) the text would be converted to a link
"#1234" w/o the leading zero when rendering the issue on the web.
This was a problem when a user didn't mean to refer to an issue
numbered 1234, but was using it in some other environment-specific
context.